Hi Aris, I put the script in beaker as /kernel/cgroups/stress tests. I only put the stress_cgroup_devices.sh file as you have suggested. Example run is:
https://beaker.engineering.redhat.com/jobs/361837

it fails for rhel5 accordingly since cgroups don't work in it, and rhel6 passes
, rhel7 , well it didn't even install so i cloned it.

Even though it's beakerized, i have made a patch to make some beaker-agnostic
changes to the script. I am going to attach a git diff to this email if you'd
like to incorporate them.

Thanks,
Gurhan